backend/requirements.txt
flask==2.3.3
werkzeug==2.3.7
requests==2.31.0
pdfplumber==0.10.3
reportlab==4.1.0
pytesseract==0.3.10
Pillow==10.2.0 